requirements.txt
anyio==4.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
cryptography==42.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
paramiko==3.4.0
pycparser==2.21
PyNaCl==1.5.0
python-decouple==3.8
python-telegram-bot==20.8
sniffio==1.3.0
